the end

+++ exercises/23_conversions/using_as.rs
@@ -10,11 +10,10 @@
 // Execute `rustlings hint using_as` or use the `hint` watch subcommand for a
 // hint.
 
-// I AM NOT DONE
 
 fn average(values: &[f64]) -> f64 {
     let total = values.iter().sum::<f64>();
-    total / values.len()
+    total / values.len() as f64
 }
 
 fn main() {
@@ -31,3 +30,7 @@ mod tests {
         assert_eq!(average(&[3.5, 0.3, 13.0, 11.7]), 7.125);
     }
 }
+
+// ! Writeup !
+// Manque de la conversion de values.len() en f64 pour que la division soit possible,
+// alors j'ai ajouté un "as f64" après values.len() dans la fonction average()
